dix: move inpututils.h and rename it to inpututil_privs.h
it's a private header, thus should not be in public include directory.
