+++ Xext/xtest.c
@@ -38,6 +38,7 @@
 #include "dix/input_priv.h"
 #include "dix/dix_priv.h"
 #include "dix/exevents_priv.h"
+#include "dix/inpututils_priv.h"
 #include "mi/mi_priv.h"
 #include "mi/mipointer_priv.h"
 #include "miext/extinit_priv.h"
@@ -60,7 +61,6 @@
 #include "mipointer.h"
 #include "xserver-properties.h"
 #include "eventstr.h"
-#include "inpututils.h"
 
 Bool noTestExtensions = FALSE;
