+++ Xi/xigrabdev.c
@@ -36,13 +36,13 @@
 
 #include "dix/dix_priv.h"
 #include "dix/exevents_priv.h"
+#include "dix/inpututils_priv.h"
 #include "dix/resource_priv.h"
 
 #include "inputstr.h"           /* DeviceIntPtr      */
 #include "windowstr.h"          /* window structure  */
 #include "exglobals.h"          /* BadDevice */
 #include "xigrabdev.h"
-#include "inpututils.h"
 
 int _X_COLD
 SProcXIGrabDevice(ClientPtr client)
