+++ Xi/xipassivegrab.c
@@ -37,6 +37,7 @@
 #include "dix/dix_priv.h"
 #include "dix/dixgrabs_priv.h"
 #include "dix/exevents_priv.h"
+#include "dix/inpututils_priv.h"
 
 #include "inputstr.h"           /* DeviceIntPtr      */
 #include "windowstr.h"          /* window structure  */
@@ -44,7 +45,6 @@
 #include "exglobals.h"          /* BadDevice */
 #include "xipassivegrab.h"
 #include "misc.h"
-#include "inpututils.h"
 
 int _X_COLD
 SProcXIPassiveGrabDevice(ClientPtr client)
