+++ dix/events.c
@@ -125,6 +125,7 @@ Equipment Corporation.
 #include "dix/exevents_priv.h"
 #include "dix/extension_priv.h"
 #include "dix/input_priv.h"
+#include "dix/inpututils_priv.h"
 #include "dix/reqhandlers_priv.h"
 #include "dix/resource_priv.h"
 #include "dix/window_priv.h"
@@ -141,7 +142,6 @@ Equipment Corporation.
 #include "resource.h"
 #include "windowstr.h"
 #include "inputstr.h"
-#include "inpututils.h"
 #include "scrnintstr.h"
 #include "cursorstr.h"
 #include "dixstruct.h"
