+++ dix/gestures.c
@@ -29,6 +29,7 @@
 #include "dix/dixgrabs_priv.h"
 #include "dix/eventconvert.h"
 #include "dix/input_priv.h"
+#include "dix/inpututils_priv.h"
 #include "dix/resource_priv.h"
 #include "mi/mi_priv.h"
 #include "os/bug_priv.h"
@@ -38,7 +39,6 @@
 #include "eventstr.h"
 #include "exevents.h"
 #include "exglobals.h"
-#include "inpututils.h"
 #include "windowstr.h"
 
 #define GESTURE_HISTORY_SIZE 100
