+++ hw/xquartz/darwinEvents.c
@@ -46,6 +46,7 @@
 #include <X11/Xmd.h>
 #include <X11/Xproto.h>
 
+#include "dix/inpututils_priv.h"
 #include "mi/mi_priv.h"
 #include "os/client_priv.h"
 
@@ -53,7 +54,6 @@
 #include "windowstr.h"
 #include "pixmapstr.h"
 #include "inputstr.h"
-#include "inpututils.h"
 #include "eventstr.h"
 #include "scrnintstr.h"
 #include "mipointer.h"
