+++ mi/mieq.c
@@ -44,6 +44,7 @@ in this Software without prior written authorization from The Open Group.
 #include   "dix/cursor_priv.h"
 #include   "dix/dix_priv.h"
 #include   "dix/input_priv.h"
+#include   "dix/inpututils_priv.h"
 #include   "mi/mi_priv.h"
 #include   "mi/mipointer_priv.h"
 #include   "os/bug_priv.h"
@@ -53,7 +54,6 @@ in this Software without prior written authorization from The Open Group.
 #include   "windowstr.h"
 #include   "pixmapstr.h"
 #include   "inputstr.h"
-#include   "inpututils.h"
 #include   "mipointer.h"
 #include   "scrnintstr.h"
 #include   "exglobals.h"
