+++ test/xi2/protocol-eventconvert.c
@@ -30,10 +30,10 @@
 
 #include "dix/exevents_priv.h"
 #include "dix/eventconvert.h"
+#include "dix/inpututils_priv.h"
 
 #include "inputstr.h"
 #include "eventstr.h"
-#include "inpututils.h"
 
 #include "protocol-common.h"
