+++ xkb/xkbAccessX.c
@@ -37,13 +37,13 @@ THE USE OR PERFORMANCE OF THIS SOFTWARE.
 #include <X11/extensions/XIproto.h>
 
 #include "dix/input_priv.h"
+#include "dix/inpututils_priv.h"
 #include "os/log_priv.h"
 #include "xkb/xkbsrv_priv.h"
 
 #include "exglobals.h"
 #include "inputstr.h"
 #include "eventstr.h"
-#include "inpututils.h"
 
 int XkbDfltRepeatDelay = 660;
 int XkbDfltRepeatInterval = 40;
